Create Lambda-Role-JSON
Attach this JSON text into a Lambda role and assign that role to your Lambda function. This will allow Lambda to create a log group based on that function and helped me with troubleshooting my function as it did not work immediately. Along with this policy you should also allow Lambda full CloudTrail access so that it is able to enable CloudTrail trails.