Add files via upload
This documentation goes through the steps in setting up automation with Lambda and CloudWatch Events to automatically detect when a CloudTrail trail is disabled. The Lambda function will automatically re enable that trail and send you an SNS notification so you can dive deeper into who was being malicious and why.